fix(intw_utility): Use nint instead of int to calculate the number of points per stage in generate_and_allocate_kpath.
Using int produces 0 in some special cases.

+++ src/modules/utility.f90
@@ -532,7 +532,7 @@ end function intgr_spline_gaussq
 
     ! number of points in the path per stage
     do i=2,nkspecial
-        nkstage(i-1) = int( real(nkpath-1,dp) * lstage(i-1) / lpath)
+        nkstage(i-1) = nint( real(nkpath-1,dp) * lstage(i-1) / lpath)
     end do
 
     ! check how many point will we actually generate
